Rename executor.rs to system_executor.rs and expose SystemExecutor
No logical changes.
